requirements_txt.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Training & Evaluation
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Model Analysis
thop>=0.1.1

# Utilities
pyyaml>=6.0
tensorboard>=2.13.0

# Optional (for advanced features)
# albumentations>=1.3.0  # Advanced augmentation
# timm>=0.9.0  # Pretrained models
# wandb>=0.15.0  # Experiment tracking